wrote some code
