+++ xmitgcm/test/test_mds_store.py
@@ -294,11 +294,11 @@ def test_read_mds_no_meta(all_mds_datadirs):
             assert isinstance(res[prefix], dask.array.core.Array)
             assert res[prefix].shape == shape
 
-
-def test_read_raw_data_llc(llc_mds_datadirs):
+@pytest.mark.parametrize("method", ["smallchunks", "bigchunks"])
+def test_read_raw_data_llc(llc_mds_datadirs, method):
     dirname, expected = llc_mds_datadirs
 
-    from xmitgcm.llc_utils import read_3d_llc_data
+    from xmitgcm.utils import read_3d_llc_data
 
     shape = expected['shape']
     nz, nface, ny, nx = shape
@@ -306,14 +306,15 @@ def test_read_raw_data_llc(llc_mds_datadirs):
     dtype = expected['dtype'].newbyteorder('>')
 
     # if we use memmap=True, we open too many files
+    kwargs = dict(method=method, dtype=dtype, memmap=False)
 
     fname = os.path.join(dirname, 'T.%010d.data' % expected['test_iternum'])
-    data = read_3d_llc_data(fname, nz, nx, dtype=dtype, memmap=False)
+    data = read_3d_llc_data(fname, nz, nx, **kwargs)
     assert data.shape == shape
     assert data.compute().shape == shape
 
     fname = os.path.join(dirname, 'XC.data')
-    data = read_3d_llc_data(fname, 1, nx, dtype=dtype, memmap=False)
+    data = read_3d_llc_data(fname, 1, nx, **kwargs)
     # make sure the first dimension is squeezed off
     assert data.shape == shape[1:]
     assert data.compute().shape == shape[1:]
@@ -490,7 +491,7 @@ def test_swap_dims(all_mds_datadirs):
         print(ds)
         assert 'XC' in ds['S'].dims
         assert 'YC' in ds['S'].dims
- 
+
 
 
 def test_prefixes(all_mds_datadirs):
